backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu==2.16.1
pillow
numpy
pandas
requests
python-multipart
aiofiles
